Updated ReceptionistView to make use of the SqlDatabaseAccess class. Updated the UpdateBalance class to use a decimal instead of an int for cost.

+++ Project-2-EMS/Models/DatabaseModels/SqlQueryModels/PatientInfoQueryModels/UpdateBalanceBy_Cost_PatientId.cs
@@ -3,10 +3,10 @@ using System.Data.SqlClient;
 
 namespace Project_2_EMS.Models.DatabaseModels {
     public class UpdateBalanceBy_Cost_PatientId : ISqlQuery, INonQuery {
-        private readonly int _Cost;
+        private readonly decimal _Cost;
         private readonly int _PatientId;
 
-        public UpdateBalanceBy_Cost_PatientId(int cost, int patientId) {
+        public UpdateBalanceBy_Cost_PatientId(decimal cost, int patientId) {
             _Cost = cost;
             _PatientId = patientId;
         }
